Moving the ScaleLineAutomationPeer class
- Removed the previously added logic to roundoff the map scale
- Added a method GetRoundedValueForAutomationPeer which will handle the rounding of MapScale for the Narrator
- Reused part of GetRoundedValue logic in GetRoundedValueForAutomationPeer. So, marked the GetRoundedValue method as internal
- Moved the ScaleLineAutomationPeer  class out of ScaleLine Class
